Brick.BChan

Documentation

data BChan a

BChan is an abstract type representing a bounded FIFO channel.

newBChan

Arguments

:: Int
maximum number of elements the channel can hold

-> IO (BChan a)

Builds and returns a new instance of BChan.

writeBChan :: BChan a -> a -> IO ()

Writes a value to a BChan; blocks if the channel is full.

writeBChanNonBlocking :: BChan a -> a -> IO Bool

Attempts to write a value to a BChan. If the channel has room, the value is written and this returns True. Otherwise this returns False and returns immediately.

readBChan :: BChan a -> IO a

Reads the next value from the BChan; blocks if necessary.

readBChan2 :: BChan a -> BChan b -> IO (Either a b)

Reads the next value from either BChan, prioritizing the first BChan; blocks if necessary.