Brick.BorderMap

Documentation

data BorderMap a

A BorderMap a is like a Map Location a, except that there is a rectangle, and only Locations on the border of this rectangle are retained. The BorderMap can be queried for the position and size of the rectangle. There are also efficient bulk query and bulk update operations for adjacent positions on the border.

data Edges a

Constructors

Edges

Fields

eTop, eBottom, eLeft, eRight :: a

eTopL :: forall a. Lens' (Edges a) a

eBottomL :: forall a. Lens' (Edges a) a

eRightL :: forall a. Lens' (Edges a) a

eLeftL :: forall a. Lens' (Edges a) a

empty :: BorderMap a

An empty BorderMap that only tracks the point (0,0).

emptyCoordinates :: Edges Int -> BorderMap a

Given a rectangle (specified as the coordinates of the top, left, bottom, and right sides), initialize an empty BorderMap.

singleton :: Location -> a -> BorderMap a

A BorderMap that tracks only the given the point (and initially maps it to the given value).

insertH :: Location -> Run a -> BorderMap a -> BorderMap a

Bulk insertion of horizontally-adjacent values. The Location gives the start point, and the Run extends in the "larger columns" direction.

insertV :: Location -> Run a -> BorderMap a -> BorderMap a

Bulk insertion of vertically-adjacent values. The Location gives the start point, and the Run extends in the "larger rows" direction.

insert :: Location -> a -> BorderMap a -> BorderMap a

Insert a single value at the given location.

unsafeUnion :: BorderMap a -> BorderMap a -> BorderMap a

Assumes the two BorderMaps are tracking the same rectangles, but have disjoint keys. This property is not checked.

coordinates :: BorderMap a -> Edges Int

The positions of the edges of the rectangle whose border is retained in a BorderMap. For example, if coordinates m = e, then the top border contains the Locations on row eTop e and between columns eLeft e to eRight e inclusive.

bounds :: BorderMap a -> Edges (Int, Int)

A complementary way to query the edges of the rectangle whose border is retained in a BorderMap. For example, if bounds m = b, then a Location's column must be between fst (eTop b) and snd (eTop b) to be retained. See also coordinates, which is in most cases a more natural border query.

values :: BorderMap a -> Edges (IMap a)

Maps giving the values along each edge. Corner values are replicated in all relevant edges.

lookupRow :: Int -> BorderMap a -> IMap a

Look up all values on a given row. The IMap returned maps columns to values.

lookupCol :: Int -> BorderMap a -> IMap a

Look up all values on a given column. The IMap returned maps rows to values.

lookupH :: Location -> Run ignored -> BorderMap a -> IMap a

Bulk lookup of horizontally-adjacent values. The Location gives the starting point, and the Run extends in the "larger columns" direction. The IMap returned maps columns to values.

lookupV :: Location -> Run ignored -> BorderMap a -> IMap a

Bulk lookup of vertically-adjacent values. The Location gives the starting point, and the Run extends in the "larger rows" direction. The IMap returned maps rows to values.

lookup :: Location -> BorderMap a -> Maybe a

Look up a single position.

setCoordinates :: Edges Int -> BorderMap a -> BorderMap a

Set the rectangle being tracked by this BorderMap, throwing away any values that do not lie on this new rectangle.

crop :: Edges Int -> BorderMap a -> BorderMap a

Ensure that the rectangle being tracked by this BorderMap extends no farther than the given one.

expand :: Edges Int -> BorderMap a -> BorderMap a

Ensure that the rectangle being tracked by this BorderMap extends at least as far as the given one.

translate :: Location -> BorderMap a -> BorderMap a

Move a BorderMap by adding the given Location to all keys in the map.