Brick.Types

Description

Basic types used by this library.

The Widget type

data Widget n

The type of widgets.

Constructors

Widget

Fields

hSize :: Size

This widget's horizontal growth policy

vSize :: Size

This widget's vertical growth policy

render :: RenderM n (Result n)

This widget's rendering function

Location types and lenses

data Location

A terminal screen location.

Constructors

Location

Fields

loc :: (Int, Int)

(Column, Row)

locL :: Lens' Location (Int, Int)

class TerminalLocation a where

The class of types that behave like terminal locations.

Methods

locationColumnL :: Lens' a Int

Get the column out of the value

locationColumn :: a -> Int

locationRowL :: Lens' a Int

Get the row out of the value

locationRow :: a -> Int

data CursorLocation n

A cursor location. These are returned by the rendering process.

Constructors

CursorLocation

Fields

cursorLocation :: !Location

The location

cursorLocationName :: !(Maybe n)

The name of the widget associated with the location

cursorLocationL :: forall n. Lens' (CursorLocation n) Location

cursorLocationNameL :: forall n n. Lens (CursorLocation n) (CursorLocation n) (Maybe n) (Maybe n)

Viewports

data Viewport

Describes the state of a viewport as it appears as its most recent rendering.

Constructors

VP

Fields

_vpLeft :: Int

The column offset of left side of the viewport.

_vpTop :: Int

The row offset of the top of the viewport.

_vpSize :: DisplayRegion

The size of the viewport.

data ViewportType

The type of viewports that indicates the direction(s) in which a viewport is scrollable.

Constructors

Vertical
Viewports of this type are scrollable only vertically.

Horizontal
Viewports of this type are scrollable only horizontally.

Both
Viewports of this type are scrollable vertically and horizontally.

vpSize :: Lens' Viewport DisplayRegion

vpTop :: Lens' Viewport Int

vpLeft :: Lens' Viewport Int

Event-handling types

newtype EventM n a

The monad in which event handlers run. Although it may be tempting to dig into the reader value yourself, just use lookupViewport.

Constructors

EventM

Fields

runEventM :: ReaderT (EventRO n) (StateT (EventState n) IO) a

data Next a

The type of actions to take upon completion of an event handler.

data BrickEvent n e

The type of events.

Constructors

VtyEvent Event
The event was a Vty event.

AppEvent e
The event was an application event.

MouseDown n Button [Modifier] Location
A mouse-down event on the specified region was received. The n value is the resource name of the clicked widget (see clickable).

MouseUp n (Maybe Button) Location
A mouse-up event on the specified region was received. The n value is the resource name of the clicked widget (see clickable).

handleEventLensed

Arguments

:: a
The state value.

-> Lens' a b
The lens to use to extract and store the target of the event.

-> (e -> b -> EventM n b)
The event handler.

-> e
The event to handle.

-> EventM n a

A convenience function for handling events intended for values that are targets of lenses in your application state. This function obtains the target value of the specified lens, invokes handleEvent on it, and stores the resulting transformed value back in the state using the lens.

Rendering infrastructure

type RenderM n a = ReaderT Context (State (RenderState n)) a

The type of the rendering monad. This monad is used by the library's rendering routines to manage rendering state and communicate rendering parameters to widgets' rendering functions.

getContext :: RenderM n Context

Get the current rendering context.

The rendering context

data Context

The rendering context. This tells widgets how to render: how much space they have in which to render, which attribute they should use to render, which bordering style should be used, and the attribute map available for rendering.

attrL :: forall r. Getting r Context Attr

The rendering context's current drawing attribute.

availWidthL :: Lens' Context Int

availHeightL :: Lens' Context Int

windowWidthL :: Lens' Context Int

windowHeightL :: Lens' Context Int

ctxAttrMapL :: Lens' Context AttrMap

ctxAttrNameL :: Lens' Context AttrName

ctxBorderStyleL :: Lens' Context BorderStyle

ctxDynBordersL :: Lens' Context Bool

Rendering results

data Result n

The type of result returned by a widget's rendering function. The result provides the image, cursor positions, and visibility requests that resulted from the rendering process.

Constructors

Result

Fields

image :: Image

The final rendered image for a widget

cursors :: [CursorLocation n]

The list of reported cursor positions for the application to choose from

visibilityRequests :: [VisibilityRequest]

The list of visibility requests made by widgets rendered while rendering this one (used by viewports)

extents :: [Extent n]

borders :: BorderMap DynBorder

Places where we may rewrite the edge of the image when placing this widget next to another one.

emptyResult :: Result n

lookupAttrName :: AttrName -> RenderM n Attr

Given an attribute name, obtain the attribute for the attribute name by consulting the context's attribute map.

data Extent n

An extent of a named area: its size, location, and origin.

Constructors

Extent

Fields

extentName :: n

extentUpperLeft :: Location

extentSize :: (Int, Int)

extentOffset :: Location

Rendering result lenses

imageL :: forall n. Lens' (Result n) Image

cursorsL :: forall n. Lens' (Result n) [CursorLocation n]

visibilityRequestsL :: forall n. Lens' (Result n) [VisibilityRequest]

extentsL :: forall n. Lens' (Result n) [Extent n]

Visibility requests

data VisibilityRequest

Constructors

VR

Fields

vrPosition :: Location

vrSize :: DisplayRegion

vrPositionL :: Lens' VisibilityRequest Location

vrSizeL :: Lens' VisibilityRequest DisplayRegion

Making lenses

suffixLenses :: Name -> DecsQ

A template haskell function to build lenses for a record type. This function differs from the makeLenses function in that it does not require the record fields to be prefixed with underscores and it adds an L suffix to lens names to make it clear that they are lenses.

Dynamic borders

bordersL :: forall n. Lens' (Result n) (BorderMap DynBorder)

data DynBorder

Information about how to redraw a dynamic border character when it abuts another dynamic border character.

Constructors

DynBorder

Fields

dbStyle :: BorderStyle

The Chars to use when redrawing the border. Also used to filter connections: only dynamic borders with equal BorderStyles will connect to each other.

dbAttr :: Attr

What Attr to use to redraw the border character. Also used to filter connections: only dynamic borders with equal Attrs will connect to each other.

dbSegments :: Edges BorderSegment

dbStyleL :: Lens' DynBorder BorderStyle

dbAttrL :: Lens' DynBorder Attr

dbSegmentsL :: Lens' DynBorder (Edges BorderSegment)

data BorderSegment

A border character has four segments, one extending in each direction (horizontally and vertically) from the center of the character.

Constructors

BorderSegment

Fields

bsAccept :: Bool

Would this segment be willing to be drawn if a neighbor wanted to connect to it?

bsOffer :: Bool

Does this segment want to connect to its neighbor?

bsDraw :: Bool

Should this segment be represented visually?

bsAcceptL :: Lens' BorderSegment Bool

bsOfferL :: Lens' BorderSegment Bool

bsDrawL :: Lens' BorderSegment Bool

data Edges a

Constructors

Edges

Fields

eTop, eBottom, eLeft, eRight :: a

eTopL :: forall a. Lens' (Edges a) a

eBottomL :: forall a. Lens' (Edges a) a

eRightL :: forall a. Lens' (Edges a) a

eLeftL :: forall a. Lens' (Edges a) a

Miscellaneous

data Size

Widget size policies. These policies communicate how a widget uses space when being rendered. These policies influence rendering order and space allocation in the box layout algorithm for hBox and vBox.

Constructors

Fixed
Widgets advertising this size policy should take up the same amount of space no matter how much they are given, i.e. their size depends on their contents alone rather than on the size of the rendering area.

Greedy
Widgets advertising this size policy must take up all the space they are given.

data Padding

The type of padding.

Constructors

Pad Int
Pad by the specified number of rows or columns.

Max
Pad up to the number of available rows or columns.

data Direction

Scrolling direction.

Constructors

Up
Up/left

Down
Down/right

Renderer internals (for benchmarking)

data RenderState n

Orphan instances

TerminalLocation (CursorLocation n)

Instance details

Methods

locationColumnL :: Lens' (CursorLocation n) Int

locationColumn :: CursorLocation n -> Int

locationRowL :: Lens' (CursorLocation n) Int

locationRow :: CursorLocation n -> Int