Brick.Util

Description

Utility functions.

Documentation

clamp

Arguments

:: Ord a

=> a
The minimum value

-> a
The maximum value

-> a
The value to clamp

-> a

Given a minimum value and a maximum value, clamp a value to that range (values less than the minimum map to the minimum and values greater than the maximum map to the maximum).

>>> clamp 1 10 11
10
>>> clamp 1 10 2
2
>>> clamp 5 10 1
5

on

Arguments

:: Color
The foreground color

-> Color
The background color

-> Attr

Build an attribute from a foreground color and a background color. Intended to be used infix.

fg :: Color -> Attr

Create an attribute from the specified foreground color (the background color is the "default").

bg :: Color -> Attr

Create an attribute from the specified background color (the background color is the "default").

clOffset :: CursorLocation n -> Location -> CursorLocation n

Add a Location offset to the specified CursorLocation.