Brick.Widgets.ProgressBar

Description

This module provides a progress bar widget.

Documentation

progressBar

Arguments

:: Maybe String
The label. If specified, this is shown in the center of the progress bar.

-> Float
The progress value. Should be between 0 and 1 inclusive.

-> Widget n

Draw a progress bar with the specified (optional) label and progress value. This fills available horizontal space and is one row high.

Attributes

progressCompleteAttr :: AttrName

The attribute of the completed portion of the progress bar.

progressIncompleteAttr :: AttrName

The attribute of the incomplete portion of the progress bar.