Brick3D.Rasterization

Documentation

type PixelAttr = (Char, Attr)

Represents one Pixel

mergeAttr :: Map (Int, Int) (Float, PixelAttr) -> Map (Int, Int) (Float, PixelAttr) -> Map (Int, Int) (Float, PixelAttr)

Merge two Map of Pixels into one by comparing zBuffer

toCanvasPixels :: Map (Int, Int) (Char, Attr) -> [((Int, Int), Char, Attr)]

Convert Map to list so that canvasSetMany can treat

rasterizeMany :: (Foldable t, Functor t) => (Int, Int) -> t DCPrimitive -> [((Int, Int), Char, Attr)]

Rasterize many of DCPrimitives

rasterize :: (Int, Int) -> DCPrimitive -> Map (Int, Int) (Float, PixelAttr)

Rasterize one DCPrimitive

rasterizeLine :: SCVertex -> SCVertex -> Vector SCVertex

DCVertexs which constructs line begin at begin and end at end

JP: 与えられた begin と end を両端に持つ線分を構成する DCVertex を返します

fillTriangle :: SCVertex -> SCVertex -> SCVertex -> Vector SCVertex

Returns DCVertexs that constructs one filled-triangle

Note that depth of those DCVertexs are not calculated properly. Currently it inherits depth of v1 I want to fix this later.

isInsideOfTri :: SCPosition -> (SCPosition, SCPosition, SCPosition) -> Bool

True if given coordinate is within given Triangle

It doesn't care if the triangle is formed well. (e.g. one vertex is on the line made of other two vertex, two of them are the same.) This is because projection could

JP: この関数は「三角形がきちんと三角形であるか」を考慮しません。 (例えば: 頂点が一直線上に並んでしまっている, 複数の頂点が同じ位置にあるなど)