Brick3D.Renderer

Documentation

render :: MonadIO m => ThreeDState -> m ThreeDState

Renders ThreeDState to one Canvas, which will be shown in Widget

render' :: ThreeDState -> [((Int, Int), Char, Attr)]

Do rendering process other than IO operation (which is done in render

farNearClip :: Camera -> Primitive -> Bool

True if given Primitive is not clipped by far/near plane

far/near平面によって描画されているかを確認する。 x,y方向の確認はしない

projectPrimitive :: Float -> Primitive -> Maybe DCPrimitive

Project one Primitive to device coordinate

projectVertex :: Float -> Vertex -> Maybe DCVertex

Project one vertex to device coordinate

applyCameraTransform :: Camera -> Primitive -> Primitive