Data.IMap

Documentation

data IMap a

Semantically, IMap and IntMap are identical; but IMap is more efficient when large sequences of contiguous keys are mapped to the same value.

data Run a

Run n a represents n copies of the value a.

Constructors

Run

Fields

len :: !Int

val :: !a

empty :: IMap a

null :: IMap a -> Bool

singleton :: Int -> Run a -> IMap a

insert :: Int -> Run a -> IMap a -> IMap a

delete :: Int -> Run ignored -> IMap a -> IMap a

restrict :: Int -> Run ignored -> IMap a -> IMap a

Given a range of keys (as specified by a starting key and a length for consistency with other functions in this module), restrict the map to keys in that range. restrict k r m is equivalent to intersectionWith const m (insert k r empty) but potentially more efficient.

lookup :: Int -> IMap a -> Maybe a

splitLE :: Int -> IMap a -> (IMap a, IMap a)

splitLE n m produces a tuple (le, gt) where le has all the associations of m where the keys are <= n and gt has all the associations of m where the keys are > n.

intersectionWith :: (a -> b -> c) -> IMap a -> IMap b -> IMap c

mapMaybe :: (a -> Maybe b) -> IMap a -> IMap b

addToKeys :: Int -> IMap a -> IMap a

Increment all keys by the given amount. This is like mapKeysMonotonic, but restricted to partially-applied addition.

unsafeUnion :: IMap a -> IMap a -> IMap a

This function is unsafe because it assumes there is no overlap between its arguments. That is, in the call unsafeUnion a b, the caller must guarantee that if lookup k a = Just v then lookup k b = Nothing and vice versa.

fromList :: [(Int, Run a)] -> IMap a

unsafeRuns :: IMap a -> IntMap (Run a)

This function is unsafe because IMaps that compare equal may split their runs into different chunks; consumers must promise that they do not treat run boundaries specially.

unsafeToAscList :: IMap a -> [(Int, Run a)]

This function is unsafe because IMaps that compare equal may split their runs into different chunks; consumers must promise that they do not treat run boundaries specially.