Brick.Markup

Description

This module provides an API for turning "markup" values into widgets. This module uses the Data.Text.Markup interface in this package to assign attributes to substrings in a text string; to manipulate markup using (for example) syntax highlighters, see that module.

Documentation

data Markup a

Markup with metadata type a assigned to each character.

markup :: (Eq a, GetAttr a) => Markup a -> Widget n

Build a widget from markup.

(@?) :: Text -> AttrName -> Markup AttrName

Build a piece of markup from text with an assigned attribute name. When the markup is rendered, the attribute name will be looked up in the rendering context's AttrMap to determine the attribute to use for this piece of text.

class GetAttr a where

A type class for types that provide access to an attribute in the rendering monad. You probably won't need to instance this.

Minimal complete definition

getAttr

Methods

getAttr :: a -> RenderM n Attr

Where to get the attribute for this attribute metadata.