Brick.Widgets.Border

Description

This module provides border widgets: vertical borders, horizontal borders, and a box border wrapper widget. All functions in this module use the rendering context's active BorderStyle; to change the BorderStyle, use withBorderStyle.

Border wrapper

border :: Widget n -> Widget n

Put a border around the specified widget.

borderWithLabel

Arguments

:: Widget n
The label widget

-> Widget n
The widget to put a border around

-> Widget n

Put a border around the specified widget with the specified label widget placed in the middle of the top horizontal border.

Horizontal border

hBorder :: Widget n

A horizontal border. Fills all horizontal space.

hBorderWithLabel

Arguments

:: Widget n
The label widget

-> Widget n

A horizontal border with a label placed in the center of the border. Fills all horizontal space.

Vertical border

vBorder :: Widget n

A vertical border. Fills all vertical space.

Drawing single border elements

borderElem :: (BorderStyle -> Char) -> Widget n

Draw the specified border element using the active border style using borderAttr.

Border attribute names

borderAttr :: AttrName

The top-level border attribute name.

vBorderAttr :: AttrName

The vertical border attribute name.

hBorderAttr :: AttrName

The horizontal border attribute name.

hBorderLabelAttr :: AttrName

The attribute used for horizontal border labels.

tlCornerAttr :: AttrName

The attribute used for border box top-left corners.

trCornerAttr :: AttrName

The attribute used for border box top-right corners.

blCornerAttr :: AttrName

The attribute used for border box bottom-left corners.

brCornerAttr :: AttrName

The attribute used for border box bottom-right corners.