Brick.Widgets.Border.Style

Description

This module provides styles for borders as used in terminal applications. Your mileage may vary on some of the fancier styles due to varying support for some border characters in the fonts your users may be using. Because of this, we provide the ascii style in addition to the Unicode styles. The unicode style is also a safe bet.

To use these in your widgets, see withBorderStyle. By default, widgets rendered without a specified border style use unicode via the Default instance provided by BorderStyle.

Documentation

data BorderStyle

A border style for use in any widget that needs to render borders in a consistent style.

Constructors

BorderStyle

Fields

bsCornerTL :: Char

Top-left corner character

bsCornerTR :: Char

Top-right corner character

bsCornerBR :: Char

Bottom-right corner character

bsCornerBL :: Char

Bottom-left corner character

bsIntersectFull :: Char

Full intersection (cross)

bsIntersectL :: Char

Left side of a horizontal border intersecting a vertical one

bsIntersectR :: Char

Right side of a horizontal border intersecting a vertical one

bsIntersectT :: Char

Top of a vertical border intersecting a horizontal one

bsIntersectB :: Char

Bottom of a vertical border intersecting a horizontal one

bsHorizontal :: Char

Horizontal border character

bsVertical :: Char

Vertical border character

borderStyleFromChar :: Char -> BorderStyle

Make a border style using the specified character everywhere.

ascii :: BorderStyle

An ASCII border style which will work in any terminal.

unicode :: BorderStyle

A unicode border style with real corner and intersection characters.

unicodeBold :: BorderStyle

A unicode border style in a bold typeface.

unicodeRounded :: BorderStyle

A unicode border style with rounded corners.

defaultBorderStyle :: BorderStyle